re-update of the font script version phone
so the github site work now wiith the font. All that was needed was a script prrovided by the site when i signed up. I understand but i dont like it...for the next project ill see if i use font awesome or not, i dont want to be dependant o a site.