check almost ok
the check animation is almost ok. The key was to use the overflow x and y and set them to hidden. But !, but, there is always a but or else its no fun, there is now a small difference between the div and the check, maybe because of the shadow. Ill check this out.